Move profile domain classes into '.profile' sub-package
